Disallow attribute definition retrieval in DB engine.
Since Winter uses behaviours to attach Scout functionality, there's no way to define attributes for the methods queried, as these methods form part of the behaviour.

We could potentially re-route the attributes somewhere else in the model at a later stage, but since this is a DB only feature, we'll leave it for now.